10.2版本

+++ src/main/java/com/dgut/cloud_disk/controller/DirectoryFileController.java
@@ -1,10 +1,15 @@
 package com.dgut.cloud_disk.controller;
 
 import com.alibaba.fastjson.JSONObject;
+import com.dgut.cloud_disk.mapper.DirectoryFileMapper;
+import com.dgut.cloud_disk.mapper.DirectoryMapper;
 import com.dgut.cloud_disk.pojo.CdstorageUser;
+import com.dgut.cloud_disk.pojo.Directory;
 import com.dgut.cloud_disk.pojo.DirectoryFile;
 import com.dgut.cloud_disk.pojo.Toshare;
+import com.dgut.cloud_disk.service.DepartmentUserService;
 import com.dgut.cloud_disk.service.DirectoryFileService;
+import com.dgut.cloud_disk.service.DirectoryService;
 import com.dgut.cloud_disk.util.JSONResult;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -26,7 +31,12 @@ import java.util.UUID;
 public class DirectoryFileController {
     @Resource
     private DirectoryFileService DFService;
-
+    @Resource
+    private DirectoryMapper Dmapper;
+    @Resource
+    private DirectoryFileMapper DFmapper;
+    @Resource
+    private DepartmentUserService DUService;
     @Autowired
     private JedisPool jedisPool;
     @RequestMapping("/df")
@@ -44,10 +54,6 @@ public class DirectoryFileController {
         String tokenValue = jedis.get(token);
         ObjectMapper objectMapper=new ObjectMapper();
         CdstorageUser cdstorageUser = objectMapper.readValue(tokenValue, CdstorageUser.class);
-
-
-    @RequestMapping("/user/delete")
-    public JSONResult deleteDirectoryFile(@RequestBody JSONObject jsonObject){
 
         String directID=jsonObject.getString("directID");
         String fileID=jsonObject.getString("fileID");
@@ -76,15 +82,33 @@ public class DirectoryFileController {
     }
 
 
-    @RequestMapping("/user/deleteDorDF")//彻底删除
+    @RequestMapping("/user/deleteDorDF")//彻底删除(群组中的删除)
     public JSONResult deleteDirectoryFileOrDirectory(@RequestBody JSONObject jsonObject){
+        //1-文件夹，2-文件
         int type =jsonObject.getInteger("type");
         String id =jsonObject.getString("id");
-        //1-文件夹，2-文件
-        if(DFService.deleteDorDF(type,id)){
-            return new JSONResult(200,"删除成功！","");
-        }else {
-            return new JSONResult(500,"删除失败！","");
+
+        Directory directory=new Directory();
+        if(type==1){
+            directory=Dmapper.selectByPrimaryKey(id);
+        }else if(type==2){
+            DirectoryFile directoryFile=DFmapper.selectByPrimaryKey(id);
+            directory=Dmapper.selectByPrimaryKey(directoryFile.getDfDirectId());
+        }else{
+            return new JSONResult(500,"类型错误！","");
+        }
+
+        String dePermission=DUService.selectdepartPermissionByid(directory.getDirectBelongDepart());
+        String p=dePermission.substring(2,3);//截取第三位（即删除文件夹的权限）
+
+        if(p.equals("1")) {//判断为1说明有该权限
+            if (DFService.deleteDorDF(type, id)) {
+                return new JSONResult(200, "删除成功！", "");
+            } else {
+                return new JSONResult(500, "删除失败！", "");
+            }
+        }else{
+            return new JSONResult(500, "权限不足！", "");
         }
     }
     @RequestMapping("/user/privateShare")//私密分享
@@ -134,20 +158,20 @@ public class DirectoryFileController {
         String token = jsonObject.getString("token");
         String Did = jsonObject.getString("newDirectID");
         String Fid = jsonObject.getString("fileID");
-        Integer shareTime =jsonObject.getInteger("shareTime");
-        String Code=jsonObject.getString("code");
+        Integer shareTime = jsonObject.getInteger("shareTime");
+        String Code = jsonObject.getString("code");
         //System.out.println(token+Did+Fid+shareTime+"---"+Code);
         //从token中获取分享人id
         Jedis jedis = jedisPool.getResource();
         String tokenValue = jedis.get(token);
-        ObjectMapper objectMapper=new ObjectMapper();
+        ObjectMapper objectMapper = new ObjectMapper();
         CdstorageUser cdstorageUser = objectMapper.readValue(tokenValue, CdstorageUser.class);
 
-        Toshare toshare=new Toshare();
+        Toshare toshare = new Toshare();
         toshare.setShareId(UUID.randomUUID().toString().replace("-", ""));
         toshare.setShareUserId(cdstorageUser.getUserId());
         toshare.setShareToUserId("0");//被分享人id为0，表示不指定
-        Date date=new Date();
+        Date date = new Date();
         toshare.setShareTime(date);
         toshare.setShareExpire(new Date(date.getTime() + shareTime * 1000L));
         toshare.setShareCode(Code);
@@ -155,15 +179,6 @@ public class DirectoryFileController {
             //3-外链分享文件 ！！外链分享只分享文件
             toshare.setShareType((byte) 3);
             toshare.setShareFileId(DFService.getDFidByFidAndDid(Fid,Did));//只设置ShareFileId
-
-        if(Fid!=null){
-            //1-私密分享文件 2-私密分享文件夹 3-外链分享文件 4-外链分享文件夹
-            toshare.setShareType((byte) 3);
-            toshare.setShareFileId(Fid);
-        }else if(Did!=null){
-            toshare.setShareType((byte) 4);
-            toshare.setShareDirectId(Did);
-
         }else{
             return new JSONResult(500,"分享失败！","");
         }
@@ -174,8 +189,8 @@ public class DirectoryFileController {
         return new JSONResult(200,"分享成功！",obj);
     }
 
-
-    @RequestMapping("/user/publicVerifyCode")//验证密码
+        //验证密码
+    @RequestMapping("/user/publicVerifyCode")
     public JSONResult publicVerifyCode(@RequestBody JSONObject jsonObject) {
         String shareID=jsonObject.getString("shareID");
         String code=jsonObject.getString("code");
@@ -186,10 +201,7 @@ public class DirectoryFileController {
             //需返回fileName，shareTime，downloadURL
             String fileName=DFService.getFileNameByID(shareID);
             Date shareTime=DFService.getShareTimeByID(shareID);
-
-
             //System.out.println(fileName+shareTime);
-
             String downloadUrl=DFService.fileDownload(DFService.getFileLinkByID(shareID),300L);
 
             JSONObject obj=new JSONObject();
@@ -205,5 +217,6 @@ public class DirectoryFileController {
         }
 
     }
+
 }
